Fix issue that caused the DB update to throw an exception

+++ app/code/community/Payone/Core/sql/payone_core_setup/upgrade-3.7.8-3.8.0.sql
@@ -16,4 +16,4 @@ ALTER TABLE `{{sales_flat_order_payment}}`
   ADD `payone_amz_order_reference` VARCHAR(255) NULL COMMENT 'Amazon Order Reference';
 
 ALTER TABLE `{{sales_flat_order}}`
-  ADD `payone_prevent_confirmation` TINYINT(1) UNSIGNED NULL COMMENT 'Flag to prevent confirmation mail';
+  ADD `payone_prevent_confirmation` INT(1) UNSIGNED NULL COMMENT 'Flag to prevent confirmation mail';
